requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pytesseract==0.3.10
Pillow==11.0.0
deep-translator==1.11.4
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
